+++ R/Measure.R
@@ -199,7 +199,7 @@ Measure = R6Class("Measure",
     aggregate = function(rr) {
       if (self$average == "macro") {
         aggregator = self$aggregator %??% mean
-        tab = score_measures(rr, list(self))
+        tab = score_measures(rr, list(self), reassemble = FALSE, view = rr$view)
         set_names(aggregator(tab[[self$id]]), self$id)
       } else { # "micro"
         self$score(rr$prediction(self$predict_sets))
@@ -241,6 +241,7 @@ score_single_measure = function(measure, task, learner, train_set, prediction) {
   if (is.list(prediction)) {
     ii = match(measure$predict_sets, names(prediction))
     if (anyMissing(ii)) {
+      lg$debug("Predict sets not available for measure, returning NaN", measure = measure, predict_sets = names(prediction))
       return(NaN)
     }
     prediction = do.call(c, prediction[ii])
@@ -254,7 +255,7 @@ score_single_measure = function(measure, task, learner, train_set, prediction) {
   }
 }
 
-#' @title Workhorse function to calculate a multiple scores
+#' @title Workhorse function to calculate multiple scores
 #'
 #' @description
 #' Converts `obj` from [ResampleResult] or [BenchmarkResult] to a `data.table`.
@@ -268,9 +269,10 @@ score_single_measure = function(measure, task, learner, train_set, prediction) {
 #' @return (`data.table()`) with added score columns.
 #'
 #' @noRd
-score_measures = function(obj, measures) {
-  reassemble_learners = any(map_lgl(measures, function(m) "requires_model" %in% m$properties))
-  tab = as.data.table(obj, reassemble_learners = reassemble_learners, convert_predictions = FALSE)
+score_measures = function(obj, measures, reassemble = TRUE, view = NULL) {
+  reassemble_learners = reassemble ||
+    some(measures, function(m) any(c("requires_learner", "requires_model") %in% m$properties))
+  tab = obj$data$as_data_table(view = view, reassemble_learners = reassemble_learners, convert_predictions = FALSE)
 
   for (measure in measures) {
     score = pmap_dbl(tab[, c("task", "learner", "resampling", "iteration", "prediction"), with = FALSE],