+++ R/ResampleResult.R
@@ -33,60 +33,30 @@
 #' rr$errors
 ResampleResult = R6Class("ResampleResult",
   public = list(
-    #' @field data ([data.table::data.table()])\cr
-    #' Internal data storage.
+    #' @field data (`ResultData`)\cr
+    #' Internal data storage object of type `ResultData`.
     #' We discourage users to directly work with this field.
     #' Use `as.table.table(ResampleResult)` instead.
     data = NULL,
 
-    #'
-    #' @field task ([Task])\cr
-    #' The task [resample()] operated on.
-    task = NULL,
-
-    #' @field learner ([Learner])\cr
-    #' Untrained learner [resample()] operated on.
-    learner = NULL,
-
-    #' @field resampling ([Resampling])\cr
-    #' Instantiated [Resampling] object which stores the splits into training and test.
-    resampling = NULL,
-
-    #' @field uhash (`character(1)`)\cr
-    #' Unique hash for this object.
-    uhash = NULL,
+    #' @field view (`character(1)`)\cr
+    #' Subset of uhashes in the [ResultData] object to operate on.
+    #' This field is for internal optimizations, i.e. to avoid unnecessary cloning.
+    view = NULL,
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     #' An alternative construction method is provided by [as_resample_result()].
     #'
-    #' @param task ([Task])\cr
-    #'   Single task all learners are trained on.
-    #' @param learner ([Learner])\cr
-    #'   Learner used to fit the individual models.
-    #' @param states (`list()`)\cr
-    #'   List of learner states (this includes the fitted models).
-    #' @param resampling ([Resampling])\cr
-    #'   Instantiated resampling.
-    #'   Number of iterations (`resampling$iters`) must match the number of states.
-    #' @param iterations (`integer()`)\cr
-    #'   Iteration numbers.
-    #' @param predictions (list of [Prediction])\cr
-    #'   List of prediction objects.
-    #' @param uhash (`character(1)`)\cr
-    #'   Unique hash for this `ResampleResult`. If `NULL`, a new unique hash is generated.
-    #'   This unique hash is primarily needed to group information in [BenchmarkResult]s.
-    initialize = function(task, learner, states, resampling, iterations, predictions, uhash = NULL) {
-      self$task = assert_task(task)
-      self$learner = assert_learner(learner)
-      self$resampling = assert_resampling(resampling)
-
-      self$data = data.table(
-        iteration = assert_integer(iterations, any.missing = FALSE, lower = 1L, upper = resampling$iters),
-        state = assert_list(states, len = length(iterations)),
-        prediction = assert_list(predictions, len = length(iterations))
-      )
-      self$uhash = assert_string(uhash, null.ok = TRUE) %??% UUIDgenerate()
+    #' @param data ([ResultData] | [data.table()])\cr
+    #'   An object of type [ResultData], either extracted from another [ResampleResult], another
+    #'   [BenchmarkResult], or manually constructed with [as_result_data()].
+    #' @param view (`character()`)\cr
+    #'   Single `uhash` of the [ResultData] to operate on.
+    #'   Used internally for optimizations.
+    initialize = function(data = ResultData$new(), view = NULL) {
+      self$data = assert_class(data, "ResultData")
+      self$view = assert_string(view, null.ok = TRUE)
     },
 
     #' @description
@@ -99,7 +69,7 @@ ResampleResult = R6Class("ResampleResult",
     #' Printer.
     #' @param ... (ignored).
     print = function() {
-      catf("%s of %i iterations", format(self), nrow(self$data))
+      catf("%s of %i iterations", format(self), self$data$iterations(self$view))
       catf(str_indent("* Task:", self$task$id))
       catf(str_indent("* Learner:", self$learner$id))
 
@@ -123,10 +93,10 @@ ResampleResult = R6Class("ResampleResult",
     #' with the aggregate function of the respective [Measure].
     #'
     #' @param predict_sets (`character()`)\cr
-    #'   Subset of `{"train", "test"}`.
     #' @return [Prediction].
+    #'   Subset of `{"train", "test"}`.
     prediction = function(predict_sets = "test") {
-      do.call(c, self$predictions(predict_sets = predict_sets))
+      self$data$prediction(self$view, predict_sets)
     },
 
     #' @description
@@ -137,7 +107,7 @@ ResampleResult = R6Class("ResampleResult",
     #'   Subset of `{"train", "test"}`.
     #' @return List of [Prediction] objects, one per element in `predict_sets`.
     predictions = function(predict_sets = "test") {
-      as_predictions(self$data$prediction, predict_sets)
+      self$data$predictions(self$view, predict_sets)
     },
 
     #' @description
@@ -159,12 +129,12 @@ ResampleResult = R6Class("ResampleResult",
     #'
     #' @return [data.table::data.table()].
     score = function(measures = NULL, ids = TRUE, predict_sets = "test") {
-      measures = as_measures(measures, task_type = self$task$task_type)
+      measures = as_measures(measures, task_type = self$data$task_type)
       assert_measures(measures, task = self$task, learner = self$learner)
       assert_flag(ids)
       assert_subset(predict_sets, mlr_reflections$predict_sets)
 
-      tab = score_measures(self, measures)
+      tab = score_measures(self, measures, view = self$view)
 
       if (ids) {
         set(tab, j = "task_id", value = ids(tab[["task"]]))
@@ -175,7 +145,10 @@ ResampleResult = R6Class("ResampleResult",
       }
 
       set(tab, j = "prediction", value = as_predictions(tab$prediction, predict_sets))
-      tab[]
+      cns = c("task", "task_id", "learner", "learner_id", "resampling", "resampling_id", "iteration",
+        "prediction", ids(measures))
+      cns = intersect(cns, names(tab))
+      tab[, cns, with = FALSE]
     },
 
     #' @description
@@ -185,7 +158,7 @@ ResampleResult = R6Class("ResampleResult",
     #'
     #' @return Named `numeric()`.
     aggregate = function(measures = NULL) {
-      measures = as_measures(measures, task_type = self$task$task_type)
+      measures = as_measures(measures, task_type = self$data$task_type)
       assert_measures(measures, task = self$task, learner = self$learner)
       set_names(map_dbl(measures, function(m) m$aggregate(self)), ids(measures))
     },
@@ -201,20 +174,74 @@ ResampleResult = R6Class("ResampleResult",
     #' You need to explicitly `$clone()` the object beforehand if you want to keeps
     #' the object in its previous state.
     filter = function(iters) {
-      iters = assert_integerish(iters, min.len = 1L, lower = 1L, upper = self$resampling$iters,
-        any.missing = FALSE, coerce = TRUE)
+      iters = assert_integerish(iters, lower = 1L, upper = self$resampling$iters,
+        any.missing = FALSE, unique = TRUE, coerce = TRUE)
 
-      self$data = self$data[list(unique(iters)), on = "iteration", nomatch = 0L]
+      self$data = self$data$clone(deep = TRUE)
+      fact = self$data$data$fact
+      if (!is.null(self$view)) {
+        fact = fact[list(self$view), on = "uhash", nomatch = NULL]
+      }
+
+      self$data$data$fact = fact[list(iters), on = "iteration", nomatch = NULL]
+
       invisible(self)
     }
   ),
 
   active = list(
+    #' @field task_type (`character(1)`)\cr
+    #' Task type of objects in the `ResampleResult`, e.g. `"classif"` or `"regr"`.
+    #' This is `NA` for empty [ResampleResult]s.
+    task_type = function(rhs) {
+      assert_ro_binding(rhs)
+      self$data$task_type
+    },
+
+    #' @field uhash (`character(1)`)\cr
+    #' Unique hash for this object.
+    uhash = function(rhs) {
+      assert_ro_binding(rhs)
+      uhash = self$data$uhashes(self$view)
+      if (length(uhash) == 0L) NA_character_ else uhash
+    },
+
+    #' @field task ([Task])\cr
+    #' The task [resample()] operated on.
+    task = function(rhs) {
+      assert_ro_binding(rhs)
+      tab = self$data$tasks(self$view)
+      if (nrow(tab) == 0L)
+        return(NULL)
+      tab$task[[1L]]
+    },
+
+    #' @field learner ([Learner])\cr
+    #' Learner prototype [resample()] operated on.
+    #' For a list of **trained** learners, see methods `$learners()`.
+    learner = function(rhs) {
+      assert_ro_binding(rhs)
+      tab = self$data$learners(self$view, states = FALSE)
+      if (nrow(tab) == 0L)
+        return(NULL)
+      tab$learner[[1L]]
+    },
+
+    #' @field resampling ([Resampling])\cr
+    #' Instantiated [Resampling] object which stores the splits into training and test.
+    resampling = function(rhs) {
+      assert_ro_binding(rhs)
+      tab = self$data$resamplings(self$view)
+      if (nrow(tab) == 0L)
+        return(NULL)
+      tab$resampling[[1L]]
+    },
+
     #' @field learners (list of [Learner])\cr
     #' List of trained learners, sorted by resampling iteration.
     learners = function(rhs) {
       assert_ro_binding(rhs)
-      reassemble_learner(list(self$learner), self$data$state)
+      self$data$learners(self$view)$learner
     },
 
     #' @field warnings ([data.table::data.table()])\cr
@@ -223,8 +250,7 @@ ResampleResult = R6Class("ResampleResult",
     #' Note that there can be multiple rows per resampling iteration if multiple warnings have been recorded.
     warnings = function(rhs) {
       assert_ro_binding(rhs)
-      logs = map(self$data$state, function(s) list(msg = get_log_condition(s, "warning")))
-      rbindlist(logs, idcol = "iteration", use.names = TRUE)
+      self$data$logs(self$view, "warning")
     },
 
     #' @field errors ([data.table::data.table()])\cr
@@ -233,30 +259,21 @@ ResampleResult = R6Class("ResampleResult",
     #' Note that there can be multiple rows per resampling iteration if multiple errors have been recorded.
     errors = function(rhs) {
       assert_ro_binding(rhs)
-      logs = map(self$data$state, function(s) list(msg = get_log_condition(s, "error")))
-      rbindlist(logs, idcol = "iteration", use.names = TRUE)
+      self$data$logs(self$view, "error")
     }
   ),
 
   private = list(
     deep_clone = function(name, value) {
-      if (name == "data") copy(value) else value
+      if (name == "data") value$clone(deep = TRUE) else value
     }
   )
 )
 
 #' @export
-as.data.table.ResampleResult = function(x, ..., reassemble_learners = TRUE, convert_predictions = TRUE, predict_sets = "test") { # nolint
-  assert_flag(reassemble_learners)
-  assert_flag(convert_predictions)
-
-  data.table(
-    task = list(x$task),
-    learner = if (reassemble_learners) x$learners else list(x$learner),
-    resampling = list(x$resampling),
-    iteration = x$data$iteration,
-    prediction = if (convert_predictions) x$predictions(predict_sets) else x$data$prediction
-  )
+as.data.table.ResampleResult = function(x, ..., predict_sets = "test") { # nolint
+  tab = x$data$as_data_table(view = x$view, predict_sets = predict_sets)
+  tab[, c("task", "learner", "resampling", "iteration", "prediction"), with = FALSE]
 }
 
 #' @export
@@ -283,12 +300,10 @@ as_resample_result = function(x, ...) {
 #' @rdname as_benchmark_result
 #' @export
 as_benchmark_result.ResampleResult = function(x, ...) { # nolint
-  tab = cbind(x$data,
-    data.table(
-      task = list(x$task),
-      learner = list(x$learner),
-      resampling = list(x$resampling),
-      uhash = x$uhash)
-  )
-  BenchmarkResult$new(tab)
+  rdata = x$data$clone(deep = TRUE)
+  if (!is.null(x$view)) {
+    rdata$data$fact = rdata$data$fact[list(x$view), on = "uhash", nomatch = NULL]
+    rdata$sweep()
+  }
+  BenchmarkResult$new(rdata)
 }